LZA-93: refactor YAML docment parsing (#50)
Change to parse the Document rather than the direct string contents.
This allows us to maintain the comments within the document.

This does make the code more complicated as we are stuck with the
explicit types of the yaml library. Best efforts have been completed to
limit the impact this might have on future work in this area.

+++ __tests__/main.test.ts
@@ -39,7 +39,9 @@ describe('action', () => {
   })
 
   afterEach(() => {
-    fs.unlinkSync(testFilePath)
+    if (fs.existsSync(testFilePath)) {
+      fs.unlinkSync(testFilePath)
+    }
   })
 
   it('adds three new accounts', async () => {
